+++ src/utilities.js
@@ -75,6 +75,7 @@ export function arrayBufferToDataURL(arrayBuffer, mimeType) {
   let uint8 = new Uint8Array(arrayBuffer);
 
   while (uint8.length > 0) {
+    // XXX: Babel's `toConsumableArray` helper will throw error in IE or Safari 9
     // eslint-disable-next-line prefer-spread
     chunks.push(fromCharCode.apply(null, toArray(uint8.subarray(0, chunkSize))));
     uint8 = uint8.subarray(chunkSize);